Fix : fix one philo

+++ philo/includes/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: ybaudoui <ybaudoui@student.42angoulem      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 11:17:23 by ybaudoui          #+#    #+#             */
-/*   Updated: 2022/11/21 10:35:19 by ybaudoui         ###   ########.fr       */
+/*   Updated: 2022/11/21 10:55:21 by ybaudoui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,34 +35,34 @@ typedef struct s_mutexes
 
 typedef struct s_philo_data
 {
-	int				number_of_meal;
-	int				philo_id;
-	long int		last_eat;
-	long int		death_time;
-	long int		eat_time;
-	long int		sleep_time;
-	int				philo_die;
-	long int		start_time;
-	int				philo_number;
-	int				eat_enough;
-	int				max_eat;
-	t_mutexes		*mutexes;
+	int			number_of_meal;
+	int			philo_id;
+	long int	last_eat;
+	int			death_time;
+	int			eat_time;
+	int			sleep_time;
+	int			philo_die;
+	long int	start_time;
+	int			philo_number;
+	int			eat_enough;
+	int			max_eat;
+	t_mutexes	*mutexes;
 }	t_philo_data;
 
 typedef struct s_data
 {
-	int					number_philo;
-	long int			time_to_die;
-	long int			time_to_eat;
-	long int			time_to_sleep;
-	int					all_eat;
-	long int			number_philo_must_eat;
-	long int			time_start;
-	pthread_t			*philo;
-	pthread_t			death;
-	pthread_t			fat;
-	t_philo_data		*philo_data;
-	t_mutexes			*mutexes;
+	int				number_philo;
+	int				time_to_die;
+	int				time_to_eat;
+	int				time_to_sleep;
+	int				all_eat;
+	int				number_philo_must_eat;
+	long int		time_start;
+	pthread_t		*philo;
+	pthread_t		death;
+	pthread_t		fat;
+	t_philo_data	*philo_data;
+	t_mutexes		*mutexes;
 }	t_data;
 
 /*****liB*****/
